# crypto-back-vulnerabilities-and-functionalities/src/main/resources/static/api.yaml
openapi: 3.0.0
info:
  version: v0.1.0
  title: UCrypt API
servers:
  - url: http://20.108.185.182:8888/ucrypt/api/v1
    description: test development server
components:
  parameters:
    offsetParam:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 120
        default: 20
      description: The numbers of items to return.
    sortParam:
      in: query
      name: sort
      required: false
      schema:
        type: string
      description: The sort property, if the sort property starts with `<` it will sort ASC, or `>` it will sort DESC.
    filterParam:
      in: query
      name: filter
      required: false
      schema:
        type: string
      description: Filter results using RSQL queries.
  responses:
    Error:
      description: Error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      properties:
        code:
          type: string
          description: Error code if any.
        message:
          type: string
          description: General error message
        details:
          type: string
          description: Error details.
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
        extraData:
          type: string
          description: Extra data
    Violation:
      description: Violation in API request
      discriminator:
        propertyName: type
        mapping:
          ConstraintViolation: '#/components/schemas/ConstraintViolation'
      properties:
        type:
          description: Violation type
          type: string
        message:
          description: Violation message
          type: string
    ConstraintViolation:
      allOf:
        - $ref: '#/components/schemas/Violation'
        - type: object
          properties:
            property:
              type: string
              description: Property name
...
    FileList:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contentType:
          type: string
    FileUpload:
      description: FileUpload model for creation of file
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file
...
paths:
...
  /files:
    post:
      tags:
        - Files
      summary: Create File
      operationId: createFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUpload'
      responses:
        201:
          description: Created
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Error'
    get:
      tags:
        - Files
      summary: Get List Of Files
      operationId: getFiles
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortParam'
        - $ref: '#/components/parameters/filterParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileList'
        default:
          $ref: '#/components/responses/Error'
  /files/{id}/download:
    get:
      tags:
        - Files
      summary: Download File With File Id
      operationId: downloadFile
      parameters:
        - name: id
          description: File Id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                type: string
                format: binary
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Error'
...